+++ src/main/java/net/atos/zac/flowable/delegate/PostDelegate.java
@@ -11,15 +11,19 @@ import org.flowable.common.engine.impl.el.FixedValue;
 import org.flowable.engine.delegate.DelegateExecution;
 
 public class PostDelegate extends AbstractDelegate {
-
     private static final Logger LOG = Logger.getLogger(PostDelegate.class.getName());
 
     private FixedValue template;
 
     @Override
     public void execute(final DelegateExecution delegateExecution) {
-        LOG.info("Verstuur per post besluit van zaak '%s' via %s met template '%s'."
-                .formatted(getZaakIdentificatie(delegateExecution), PostDelegate.class.getSimpleName(),
-                        template.getExpressionText()));
+        LOG.info(
+                "Verstuur per post besluit van zaak '%s' via %s met template '%s'."
+                        .formatted(
+                                getZaakIdentificatie(delegateExecution),
+                                PostDelegate.class.getSimpleName(),
+                                template.getExpressionText()
+                        )
+        );
     }
 }
