+++ src/main/java/net/atos/zac/flowable/task/exception/TaskNotFoundException.java
@@ -1,4 +1,9 @@
-package net.atos.zac.flowable.exception;
+/*
+ * SPDX-FileCopyrightText: 2024 Lifely
+ * SPDX-License-Identifier: EUPL-1.2+
+ */
+
+package net.atos.zac.flowable.task.exception;
 
 import jakarta.ws.rs.NotFoundException;
