+++ src/main/java/net/atos/zac/policy/PolicyService.java
@@ -10,7 +10,7 @@ import static net.atos.client.zgw.shared.util.URIUtil.parseUUIDFromResourceURI;
 import static net.atos.client.zgw.zrc.util.StatusTypeUtil.isHeropend;
 import static net.atos.client.zgw.zrc.util.StatusTypeUtil.isIntake;
 import static net.atos.zac.enkelvoudiginformatieobject.util.EnkelvoudigInformatieObjectCheckersKt.isSigned;
-import static net.atos.zac.flowable.TaakVariabelenService.readZaaktypeOmschrijving;
+import static net.atos.zac.flowable.task.TaakVariabelenService.readZaaktypeOmschrijving;
 import static net.atos.zac.flowable.util.TaskUtil.isOpen;
 
 import jakarta.annotation.Nullable;
