+++ src/main/java/net/atos/zac/zoeken/IndexeerService.java
@@ -47,7 +47,7 @@ import net.atos.client.zgw.zrc.ZrcClientService;
 import net.atos.client.zgw.zrc.model.Zaak;
 import net.atos.client.zgw.zrc.model.ZaakListParameters;
 import net.atos.zac.app.task.model.TaakSortering;
-import net.atos.zac.flowable.FlowableTaskService;
+import net.atos.zac.flowable.task.FlowableTaskService;
 import net.atos.zac.shared.model.SorteerRichting;
 import net.atos.zac.zoeken.converter.AbstractZoekObjectConverter;
 import net.atos.zac.zoeken.model.ZoekObject;
