+++ src/main/kotlin/net/atos/zac/app/signalering/converter/SignaleringTaskConverters.kt
@@ -1,8 +1,8 @@
 package net.atos.zac.app.signalering.converter
 
 import net.atos.zac.app.signalering.model.RestSignaleringTaskSummary
-import net.atos.zac.flowable.TaakVariabelenService.readZaakIdentificatie
-import net.atos.zac.flowable.TaakVariabelenService.readZaaktypeOmschrijving
+import net.atos.zac.flowable.task.TaakVariabelenService.readZaakIdentificatie
+import net.atos.zac.flowable.task.TaakVariabelenService.readZaaktypeOmschrijving
 import net.atos.zac.util.DateTimeConverterUtil
 import org.flowable.task.api.TaskInfo
