+++ src/main/kotlin/net/atos/zac/app/task/TaskRestService.kt
@@ -45,14 +45,14 @@ import net.atos.zac.authentication.ActiveSession
 import net.atos.zac.authentication.LoggedInUser
 import net.atos.zac.configuratie.ConfiguratieService
 import net.atos.zac.event.EventingService
-import net.atos.zac.flowable.FlowableTaskService
-import net.atos.zac.flowable.TaakVariabelenService
-import net.atos.zac.flowable.TaakVariabelenService.TAAK_DATA_DOCUMENTEN_VERZENDEN_POST
-import net.atos.zac.flowable.TaakVariabelenService.TAAK_DATA_TOELICHTING
-import net.atos.zac.flowable.TaakVariabelenService.TAAK_DATA_VERZENDDATUM
-import net.atos.zac.flowable.TaakVariabelenService.isZaakHervatten
-import net.atos.zac.flowable.TaakVariabelenService.readSignatures
-import net.atos.zac.flowable.TaakVariabelenService.readZaakUUID
+import net.atos.zac.flowable.task.FlowableTaskService
+import net.atos.zac.flowable.task.TaakVariabelenService
+import net.atos.zac.flowable.task.TaakVariabelenService.TAAK_DATA_DOCUMENTEN_VERZENDEN_POST
+import net.atos.zac.flowable.task.TaakVariabelenService.TAAK_DATA_TOELICHTING
+import net.atos.zac.flowable.task.TaakVariabelenService.TAAK_DATA_VERZENDDATUM
+import net.atos.zac.flowable.task.TaakVariabelenService.isZaakHervatten
+import net.atos.zac.flowable.task.TaakVariabelenService.readSignatures
+import net.atos.zac.flowable.task.TaakVariabelenService.readZaakUUID
 import net.atos.zac.flowable.util.TaskUtil
 import net.atos.zac.policy.PolicyService
 import net.atos.zac.policy.PolicyService.assertPolicy
