+++ src/main/kotlin/net/atos/zac/app/task/converter/RestTaskConverter.kt
@@ -12,13 +12,13 @@ import net.atos.zac.app.identity.converter.RestGroupConverter
 import net.atos.zac.app.identity.converter.RestUserConverter
 import net.atos.zac.app.policy.converter.RESTRechtenConverter
 import net.atos.zac.app.task.model.RestTask
-import net.atos.zac.flowable.TaakVariabelenService.readTaskData
-import net.atos.zac.flowable.TaakVariabelenService.readTaskDocuments
-import net.atos.zac.flowable.TaakVariabelenService.readTaskInformation
-import net.atos.zac.flowable.TaakVariabelenService.readZaakIdentificatie
-import net.atos.zac.flowable.TaakVariabelenService.readZaakUUID
-import net.atos.zac.flowable.TaakVariabelenService.readZaaktypeOmschrijving
-import net.atos.zac.flowable.TaakVariabelenService.readZaaktypeUUID
+import net.atos.zac.flowable.task.TaakVariabelenService.readTaskData
+import net.atos.zac.flowable.task.TaakVariabelenService.readTaskDocuments
+import net.atos.zac.flowable.task.TaakVariabelenService.readTaskInformation
+import net.atos.zac.flowable.task.TaakVariabelenService.readZaakIdentificatie
+import net.atos.zac.flowable.task.TaakVariabelenService.readZaakUUID
+import net.atos.zac.flowable.task.TaakVariabelenService.readZaaktypeOmschrijving
+import net.atos.zac.flowable.task.TaakVariabelenService.readZaaktypeUUID
 import net.atos.zac.flowable.util.TaskUtil
 import net.atos.zac.formulieren.FormulierDefinitieService
 import net.atos.zac.policy.PolicyService
