+++ src/main/kotlin/net/atos/zac/task/TaskService.kt
@@ -9,8 +9,8 @@ import net.atos.zac.app.task.model.RestTaskDistributeData
 import net.atos.zac.app.task.model.RestTaskReleaseData
 import net.atos.zac.authentication.LoggedInUser
 import net.atos.zac.event.EventingService
-import net.atos.zac.flowable.FlowableTaskService
-import net.atos.zac.flowable.exception.TaskNotFoundException
+import net.atos.zac.flowable.task.FlowableTaskService
+import net.atos.zac.flowable.task.exception.TaskNotFoundException
 import net.atos.zac.signalering.event.SignaleringEventUtil
 import net.atos.zac.signalering.model.SignaleringType
 import net.atos.zac.websocket.event.ScreenEventType
