+++ src/test/kotlin/net/atos/zac/app/task/TaskRestServiceTest.kt
@@ -39,10 +39,10 @@ import net.atos.zac.app.task.model.createRestTaskReleaseData
 import net.atos.zac.authentication.LoggedInUser
 import net.atos.zac.authentication.createLoggedInUser
 import net.atos.zac.event.EventingService
-import net.atos.zac.flowable.FlowableTaskService
-import net.atos.zac.flowable.TaakVariabelenService
-import net.atos.zac.flowable.TaakVariabelenService.TAAK_DATA_DOCUMENTEN_VERZENDEN_POST
-import net.atos.zac.flowable.TaakVariabelenService.TAAK_DATA_VERZENDDATUM
+import net.atos.zac.flowable.task.FlowableTaskService
+import net.atos.zac.flowable.task.TaakVariabelenService
+import net.atos.zac.flowable.task.TaakVariabelenService.TAAK_DATA_DOCUMENTEN_VERZENDEN_POST
+import net.atos.zac.flowable.task.TaakVariabelenService.TAAK_DATA_VERZENDDATUM
 import net.atos.zac.flowable.util.TaskUtil.getTaakStatus
 import net.atos.zac.identity.model.getFullName
 import net.atos.zac.policy.PolicyService
