+++ src/test/kotlin/net/atos/zac/productaanvraag/ProductaanvraagServiceTest.kt
@@ -35,8 +35,8 @@ import net.atos.zac.admin.model.createZaakafhandelParameters
 import net.atos.zac.configuratie.ConfiguratieService
 import net.atos.zac.configuratie.ConfiguratieService.BRON_ORGANISATIE
 import net.atos.zac.documenten.InboxDocumentenService
-import net.atos.zac.flowable.BPMNService
-import net.atos.zac.flowable.CMMNService
+import net.atos.zac.flowable.bpmn.BPMNService
+import net.atos.zac.flowable.cmmn.CMMNService
 import net.atos.zac.identity.IdentityService
 import net.atos.zac.productaanvraag.model.InboxProductaanvraag
 import net.atos.zac.productaanvraag.model.generated.Geometry
