+++ src/test/kotlin/net/atos/zac/task/TaskServiceTest.kt
@@ -20,8 +20,8 @@ import net.atos.zac.app.task.model.createRestTaskReleaseData
 import net.atos.zac.authentication.LoggedInUser
 import net.atos.zac.event.EventingService
 import net.atos.zac.event.Opcode
-import net.atos.zac.flowable.FlowableTaskService
-import net.atos.zac.flowable.exception.TaskNotFoundException
+import net.atos.zac.flowable.task.FlowableTaskService
+import net.atos.zac.flowable.task.exception.TaskNotFoundException
 import net.atos.zac.signalering.event.SignaleringEvent
 import net.atos.zac.websocket.event.ScreenEvent
 import net.atos.zac.websocket.event.ScreenEventType
