+++ src/test/kotlin/net/atos/zac/zoeken/converter/ZoekenConverterFixtures.kt
@@ -3,7 +3,7 @@ package net.atos.zac.zoeken.converter
 import net.atos.client.zgw.shared.ZGWApiService
 import net.atos.client.zgw.zrc.ZrcClientService
 import net.atos.client.zgw.ztc.ZtcClientService
-import net.atos.zac.flowable.FlowableTaskService
+import net.atos.zac.flowable.task.FlowableTaskService
 import net.atos.zac.identity.IdentityService
 
 @Suppress("LongParameterList")
